Change analyze
