+++ mcq-ws/src/main/java/pl/poznan/put/ws/jpa/StructureOutputCrudRepo.java
@@ -2,11 +2,11 @@ package pl.poznan.put.ws.jpa;
 
 
 import org.springframework.data.repository.CrudRepository;
-import pl.poznan.put.ws.entities.StructureOutput;
+import pl.poznan.put.ws.entities.TrigonometricRepresentation;
 
 import java.util.Optional;
 import java.util.UUID;
 
-public interface StructureOutputCrudRepo extends CrudRepository<StructureOutput, UUID> {
-    Optional<StructureOutput> findByInputId(UUID inputId);
+public interface StructureOutputCrudRepo extends CrudRepository<TrigonometricRepresentation, UUID> {
+    Optional<TrigonometricRepresentation> findByInputId(UUID inputId);
 }
