+++ mcq-ws/src/main/resources/schema/ModelComparisonDTO.yaml
